+++ inst/unit_tests/test_lin_cost.R
@@ -1,9 +1,5 @@
-library('RcppDynProg')
 
-context("test_lin_cost")
-
-
-test_that("test_lin_cost: test scoring", {
+test_lin_cost <- function() {
   y <- c(1, 2, 1, -1, 4, 5, 10)
   x <- seq_len(length(y))
   w <- 1 + numeric(length(y))
@@ -12,5 +8,8 @@ test_that("test_lin_cost: test scoring", {
   out_est <- xlin_fits_lm(x, y, w)
   c2 <- sum(w*(y-out_est)^2)
   
-  testthat::expect_true(abs(c1-c2)<=1e-3)
-})
+  RUnit::checkTrue(abs(c1-c2)<=1e-3)
+
+  invisible(NULL)
+}
+