+++ inst/unit_tests/test_logistic_cost_fn.R
@@ -1,10 +1,5 @@
-library('RcppDynProg')
 
-context("test_logistic_cost_fn")
-
-
-test_that("test_logistic_cost_fn: test scoring", {
-  
+test_logistic_cost <- function() {
   d <- data.frame(
     x = c(1, 2, 3, 4, 5, 6, 7),
     y = c(0, 0, 1, 0, 1, 1, 0)
@@ -15,5 +10,7 @@ test_that("test_logistic_cost_fn: test scoring", {
   m <- glm(y~x, data=d, family = binomial)
   dev <- summary(m)$deviance
   
-  testthat::expect_true(abs(cost-dev)<=1e-3)
-})
+  RUnit::checkTrue(abs(cost-dev)<=1e-3)
+
+  invisible(NULL)
+}