+++ NEWS.md
@@ -1,9 +1,10 @@
 
-# RcppDynProg 0.1.1 2019/02/01
+# RcppDynProg 0.1.1 2019/02/02
 
  * Add logistic scoring.
  * No k-bound variation.
  * Add a logistic fitter.
+ * Switch to RUnit testing.
 
 # RcppDynProg 0.1.0 2019/01/01